+++ src/components/Header/__tests__/Header.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import Header from '../header'
+import Header from '..'
 import { render } from '@testing-library/react'
 
 describe('<Header/>', () => {
