+++ src/components/Layout/index.tsx
@@ -8,8 +8,8 @@
 import React from 'react'
 import { useStaticQuery, graphql } from 'gatsby'
 
-import Header from '../Header/header'
-import './layout.css'
+import Header from '../Header'
+import './styles.css'
 
 const Layout: React.FC = ({ children }) => {
   const data = useStaticQuery(graphql`
