+++ src/pages/404/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import Layout from '../../components/Layout/layout'
-import SEO from '../../components/Seo/seo'
+import Layout from '../../components/Layout'
+import SEO from '../../components/Seo'
 
 const NotFoundPage = () => (
   <Layout>
