+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/Layout/layout'
-import Image from '../components/Image/image'
-import SEO from '../components/Seo/seo'
+import Layout from '../components/Layout'
+import Image from '../components/Image'
+import SEO from '../components/Seo'
 
 const IndexPage = () => (
   <Layout>
