+++ src/pages/page-2/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../../components/Layout/layout'
-import SEO from '../../components/Seo/seo'
+import Layout from '../../components/Layout'
+import SEO from '../../components/Seo'
 
 const SecondPage = () => (
   <Layout>
